finish

+++ app/Http/Controllers/ControllerProduct.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ControllerProduct extends Controller
 {
@@ -24,12 +25,13 @@ class ControllerProduct extends Controller
 
     public function store(Request $request)
     {
-        $newProducts = new Product();
-        $newProducts->name = $request->name;
-        $newProducts->price = $request->price;
-        $newProducts->description = $request->description;
-        $newProducts->save();
-        return redirect('/product');
+        $newProduct = new Product();
+        $newProduct->name = $request->name;
+        $newProduct->price = $request->price;
+        $newProduct->description = $request->description;
+        $newProduct->user_id = $request->user()->id;
+        $newProduct->save();
+        return redirect('/product')->with('operation_created', $newProduct);
     }
 
     public function details($id)
@@ -40,35 +42,46 @@ class ControllerProduct extends Controller
 
 
 
-    public function delete($id)
+    public function delete($id, Request $request)
     {
         $delete = Product::find($id);
+
+        $delete = Product::find($id);
+        if ($request->user()->id !== $delete->user_id) {
+            abort(401);
+        }
+
         $delete->delete();
-        return redirect('/product');
+        return redirect('/product')->with('operation_success', $delete);
     }
 
 
 
-    public function modify($id)
+    public function modify($id, Request $request)
     {
+        $product = Product::find($id);
+        if ($request->user()->id !== $product->user_id) {
+            abort(401);
+        }
 
         return view('modify', [
-            $product = Product::find($id),
-            'id' => $id,
-            'name' => $product->name,
-            'price' => $product->price,
-            'description' => $product->description
+            'product' => Product::findOrFail($id),
         ]);
     }
 
 
     public function update(Request $request, $id)
     {
+
+        $product = Product::find($id);
+        if ($request->user()->id !== $product->user_id) {
+            abort(401);
+        }
         $update = Product::find($id);
         $update->name = $request->name;
         $update->price = $request->price;
         $update->description = $request->description;
         $update->save();
-        return redirect('/product');
+        return redirect('/product')->with('operation_updated', $update);
     }
 }
